more integer conversion

+++ etc/test/test_versus_hinnant_tz.cpp
@@ -73,8 +73,8 @@ std::vector<entry> get_entries( string_view name, sys_seconds start, sys_seconds
             info.abbrev,
             info.begin.time_since_epoch().count(),
             info.end.time_since_epoch().count(),
-            FromUTC( info.offset.count() ),
-            RuleSave( info.save.count() * 60 ),
+            FromUTC( i32( info.offset.count() ) ),
+            RuleSave( i32( info.save.count() * 60 ) ),
         } );
         T = info.end;
     }
